+++ tests/func/test_vec_capacity.cc
@@ -60,8 +60,8 @@ int main(void)
 #else
     fprintf(stderr, "unknown libc++: __cplusplus %d\n", (int)__cplusplus);
 #endif
-    const size_t loops = TEST_RAND(TEST_MAX_LOOPS);
-    for (size_t loop = 0; loop < loops; loop++)
+    const unsigned loops = TEST_RAND(TEST_MAX_LOOPS);
+    for (unsigned loop = 0; loop < loops; loop++)
     {
         uint8_t value = TEST_RAND(UINT8_MAX); // SMALLEST SIZE.
         size_t size = loop ? TEST_RAND(TEST_MAX_SIZE) : TEST_RAND(30);
